+++ src/main/java/sa/gov/nic/bio/biokit/DeviceServiceFactory.java
@@ -4,6 +4,8 @@ import sa.gov.nic.bio.biokit.face.FaceService;
 import sa.gov.nic.bio.biokit.face.WebsocketFaceServiceImpl;
 import sa.gov.nic.bio.biokit.fingerprint.FingerprintService;
 import sa.gov.nic.bio.biokit.fingerprint.WebsocketFingerprintServiceImpl;
+import sa.gov.nic.bio.biokit.passport.PassportScannerService;
+import sa.gov.nic.bio.biokit.passport.WebsocketPassportScannerServiceImpl;
 import sa.gov.nic.bio.biokit.websocket.beans.Message;
 
 public final class DeviceServiceFactory
@@ -16,5 +18,10 @@ public final class DeviceServiceFactory
     public static FingerprintService getFingerprintService(AsyncClientProxy<Message> asyncClientProxy)
     {
         return new WebsocketFingerprintServiceImpl(asyncClientProxy);
+    }
+    
+    public static PassportScannerService getPassportScannerService(AsyncClientProxy<Message> asyncClientProxy)
+    {
+        return new WebsocketPassportScannerServiceImpl(asyncClientProxy);
     }
 }