+++ src/main/java/sa/gov/nic/bio/biokit/passport/PassportScannerService.java
@@ -1,0 +1,15 @@
+package sa.gov.nic.bio.biokit.passport;
+
+import sa.gov.nic.bio.biokit.DeviceService;
+import sa.gov.nic.bio.biokit.beans.InitializeResponse;
+import sa.gov.nic.bio.biokit.beans.ServiceResponse;
+import sa.gov.nic.bio.biokit.passport.beans.CapturePassportResponse;
+
+import java.util.concurrent.Future;
+
+public interface PassportScannerService extends DeviceService
+{
+    Future<ServiceResponse<InitializeResponse>> initialize();
+    Future<ServiceResponse<InitializeResponse>> deinitialize(String currentDeviceName);
+    Future<ServiceResponse<CapturePassportResponse>> capture(String currentDeviceName);
+}