+++ src/main/java/sa/gov/nic/bio/biokit/websocket/ServiceType.java
@@ -2,7 +2,7 @@ package sa.gov.nic.bio.biokit.websocket;
 
 public enum ServiceType
 {
-    FINGERPRINT("FINGER"), FACE("FACE");
+    FINGERPRINT("FINGER"), FACE("FACE"), PASSPORT("PASSPORT");
 
     private final String type;